==> notebooks/03_cencus_data.ipynb <==
import pandas as pd
pd.set_option('display.max_columns', None)

# Read suumo data into a pandas DataFrame
df_property = pd.read_csv('/app/data/sales_combo_property_geo.csv')
df_property.drop(["address_block"], axis=1, inplace=True)
df_property.rename(columns={'address': 'original_address'}, inplace=True)
df_property.head()

# Read 2020 cencus data into a pandas DataFrame
df_census = pd.read_csv('/app/data/tokyo_census_data_2020.csv')
df_census.head()

df_result = pd.merge(df_property, df_census, left_on='converted_address', right_on='address', how='left')
# df_result.drop(["address"], axis=1, inplace=True)
df_result.head()


print(f"{len(df_result)} {len(df_property)} {len(df_census)}")

df_result.isna().sum().to_frame(name='NaNs').T

# Store the list of column names from df_census to identify which columns to update
census_columns = df_census.columns.tolist()

# Find rows where address match was not found initially
mask = df_result['address'].isnull()

# Extract part of the address before any numbers (assuming this is your logic)
df_result['temp_address'] = df_result['original_address'].str.extract(r'(.*?)(?=\d+)', expand=False)

# Iterate through the masked rows and find matches based on the temporary modified address
for idx, row in df_result.loc[mask].iterrows():
    match = df_census[df_census['address'] == row['temp_address']]
    if not match.empty:
        # Update all columns from df_census for this row in df_result
        for col in census_columns:
            if col in df_result.columns:  # Check if the df_census column exists in df_result
                df_result.at[idx, col] = match[col].values[0]

# Drop the temporary address column after use
df_result.drop(columns=['temp_address'], inplace=True)
df_result.head()

df_result.isna().sum().to_frame(name='NaNs').T

## Drop NaN addresses

df_result = df_result.dropna(subset=['address'])

## Save

df_result.to_csv('/app/data/sales_combo_property_geo_census.csv', index=False)
